fix checkstyle / ScopedContextBuilder
