+++ pom.xml
@@ -156,6 +156,8 @@
                     <suppressionsLocation>${project.resources[0].directory}/checkstyle-ignore.xml</suppressionsLocation>
                     <consoleOutput>true</consoleOutput>
                     <failsOnError>true</failsOnError>
+                    <!-- <violationSeverity>warning</violationSeverity>
+                    <failOnViolation>true</failOnViolation> -->
                     <linkXRef>false</linkXRef>
                 </configuration>
                 <executions>
